+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # [develop](https://github.com/adhearsion/adhearsion-asr)
+
+# [v1.1.0](https://github.com/adhearsion/adhearsion-asr/compare/1.0.1...1.1.0) - [2014-01-02](https://rubygems.org/gems/adhearsion-asr/versions/1.1.0)
   * Feature: Alias `Result#response` to `#utterance`, with a deprecation warning
   * Feature: Add `#match?` predicate to `Result` object
   * Feature: When no prompts are supplied, only execute an input component